Version 1.0
Fixed a subtle concurrency bug in installing the timer handler (that was
actually a bug in DOS 2.11). Cleaned up a lot of the formatting
specifiers so numbers work properly everywhere. Added a few more sanity
checks, and a warning if the SN clock is different than the Apricot's
(which all of them will be unless there's another 2MHz SN system out
there).